# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Dazeku API",
        "description": "API per la gestione della webapp Dazeku",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "dazeku.onrender.com",
    "basePath": "/",
...
    "schemes": ["https", "http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
        "/api/user": {
            "get": {
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "cookie",
                        "name": "token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "summary": "Recupera un utente dall'id, il token come parametro nella query è necessario solo se si ha un token per autenticarsi come l'utente di cui si vogliono vedere i dati (permette di visualizzare le opere con visibilità privata) e il token non è stato salvato come cookie.",
                "responses": {
                    "200": {
                        "description": "E' stato trovato l'utente associato all'id passato come parametro e non sono avvenuti errori",                        
                        "schema": {
                            "type": "object",
                            "properties": {
                              "success": {
                                "type": "boolean"
                              },
                              "data": {
                                "$ref": "#/definitions/User"
                              }
                            },
                            "required": ["success", "data"]                            
                        }
                    },
                    "400": {
                        "description": "Parametro id non specificato",
                        "schema": {
                            "type": "object",
                            "properties": {
                              "success": {
                                "type": "boolean"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": ["success", "message"]
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "query error"
                            }
                        }
                        
                    },
                    "404": {
                        "description": "L'utente non esiste",
                        "schema": {
                            "type": "object",
                            "properties": {
                              "success": {
                                "type": "boolean"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": ["success", "message"]
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "l'utente non esiste"
                            }
                        }
                    },
                    "422": {
                        "description": "Il parametro id deve essere una stringa di 12 bytes",
                        "schema": {
                            "type": "object",
                            "properties": {
                              "success": {
                                "type": "boolean"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": ["success", "message"]
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "L'id deve essere una stringa di 12 bytes o un intero"
                            }
                        }
                        
                    },
                    "500": {
                        "description": "Il server non è riuscito a recuperare i dati dell'utente",
                        "schema": {
                            "type": "object",
                            "properties": {
                              "success": {
                                "type": "boolean"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": ["success", "message"]
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "E' avvenuto un errore nel tentativo di recuperare i dati dell'utente"
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "datatomodify",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string"
                                },
                                "nome": {
                                    "type": "string"
                                },
                                "cognome": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "token": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    {
                        "in": "cookie",
                        "name": "token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "summary": "Modifica dati utente, è necessario fornire un token di autenticazione valido, verranno modificati i parametri inviati con la richiesta, è necessario compilare solo quelli che si vuole modificare.",
                "responses": {
                    "200": {
                        "description": "Modifiche avvenute con successo.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        },
                        "examples": {
                            "application/json": {
                                "success": true,
                                "message": "modifica avvenuta con successo"
                            }
                        }
                    },
                    "400": {
                        "description": "Non sono stati specificati parametri da modificare.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "Non sono stati specificati parametri da modificare"
                            }
                        }
                    },
                    "401": {
                        "description": "token di autenticazione non fornito o non valido, il caso è specificato in 'message'",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "token non fornito"
                            }
                        }
                    },
                    "403": {
                        "description": "Lo username è già in uso.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "username non valido, eventuali altri dati sono stati modificati"
                            }
                        }
                    },
                    "500": {
                        "description": "è avvenuto un errore sul server",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "è avvenuto un errore"
                            }
                        }
                    }
                }

            },
            "delete": {
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "token",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    {
                        "in": "cookie",
                        "name": "token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "summary": "Elimina l'utente associato al token passato, se dovessi mandare due diversi token validi uno nel body e uno nel cookie viene eliminato l'utente associato al token passato nel body.\nE' necessario l'invio di un solo token.",
                "responses": {
                    "200": {
                        "description": "Utente eliminato con successo",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        },
                        "examples": {
                            "application/json": {
                                "success": true,
                                "message": "utente eliminato con successo"
                            }
                        }
                    },
                    "401": {
                        "description": "Token di autenticazione non fornito o non valido, il caso è specificato in 'message'",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "token non valido"
                            }
                        }
                    },
                    "500": {
                        "description": "E' avvenuto un errore sul server",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "è avvenuto un errore"
                            }
                        }
                    }
                }
            }
        },
        "/api/user/getmany": {
            "get": {
                "tags": [
                    "user"
                ],
                "parameters": [{
                    "in": "query",
                    "name": "username",
                    "schema": {
                        "type": "string"
                    }
                }],
                "summary": "Restituisci tutti gli utenti che contengono nello username la stringa passata come parametro",
                "responses": {
                    "200": {
                        "description": "Gli utenti sono stati trovati e restituiti con successo",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/User"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parametro username non specificato",
                        "schema": {
                            "type": "object",
                            "properties": {
                              "success": {
                                "type": "boolean"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": ["success", "message"]
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "query error"
                            }
                        }
                    },
                    "500": {
                        "description": "Il server non è riuscito a recuperare gli utenti",
                        "schema": {
                            "type": "object",
                            "properties": {
                              "success": {
                                "type": "boolean"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": ["success", "message"]
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "è avvenuto un errore"
                            }
                        }
                    }
                }
            }
        },
        "/api/user/signUp": {
            "post": {
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "userdata",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "nome": {
                                    "type": "string"
                                },
                                "cognome": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "summary": "Registra un nuovo utente",
                "responses": {
                    "200": {
                        "description": "La registrazione è avvenuta con successo",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        },
                        "examples": {
                            "application/json": {
                                "success": true,
                                "message": "registrazione avvenuta"
                            }
                        }
                    },
                    "400": {
                        "description": "Mancano uno o più parametri",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "parametri mancanti"
                            }
                        }
                    },
                    "403": {
                        "description": "La mail ha un formato non valido, oppure la mail o lo username sono già in uso, oppure la password non rispetta i requisiti di sicurezza.\n Nel primo caso message vale 'email non valida', nel secondo vale 'username o email già in uso', nel terzo vale 'la password non rispetta i requisiti di sicurezza'",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "username o email già in uso"
                            }
                        }
                    },
                    "500": {
                        "description": "E' avvenuto un errore",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "è avvenuto un errore"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "User": {
            "required": [
                "_id",
                "username",
                "nome",
                "cognome",
                "opereProdotte",
                "operePreferite"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "cognome": {
                    "type": "string"
                },
                "opereProdotte": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Opera"
                    }
                },
                "operePreferite": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Opera"
                    }
                }
            }
        },
        "Opera": {
            "required": [
                "_id",
                "titolo",
                "generi",
                "autore",
                "autoreId",
                "sinossi",
                "nSeguaci",
                "listaCap",
                "visibile"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "titolo": {
                    "type": "string"
                },
                "generi": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "autore": {
                    "type": "string"
                },
                "autoreId": {
                    "type": "string"
                },
                "sinossi": {
                    "type": "string"
                },
                "nSeguaci": {
                    "type": "number"
                },
                "listaCap": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Capitolo"                        
                    }
                },
                "visibile": {
                    "type": "boolean"
                }
            }
        },
        "Capitolo": {
            "required": [
                "titolo",
                "operaId",
                "numero",
                "testo",
                "visibile"
            ],
            "properties": {
                "titolo": {
                    "type": "string"
                },
                "operaId": {
                    "type": "string"
                },
                "numero": {
                    "type": "number"
                },
                "testo": {
                    "type": "string"
                },
                "visibile": {
                    "type": "boolean"
                }
            }
        }
    }
}